Updated mimic folder name and refactored degree helper functions to their own file
