+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,10 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\subsystem-mimic-arm",
+    "\\subsystem-mimic-arm\\implementation",
+    "\\subsystem-mimic-arm\\source"
+  ],
+  "SelectedNode": "\\subsystem-mimic-arm\\implementation\\cd74hc4067.hpp",
+  "PreviewInSolutionExplorer": false
+}
