+++ CppProperties.json
@@ -1,0 +1,22 @@
+{
+  "configurations": [
+    {
+      "inheritEnvironments": [
+        "msvc_x86"
+      ],
+      "name": "x86-Debug",
+        "includePath": [
+            "${env.INCLUDE}",
+            "${workspaceRoot}\\**",
+            "D:\\Phazonic\\Documents\\Programming\\sjsu-robotics\\SJSU-Dev2\\library\\**"
+        ],
+      "defines": [
+        "WIN32",
+        "_DEBUG",
+        "UNICODE",
+        "_UNICODE"
+      ],
+      "intelliSenseMode": "windows-msvc-x86"
+    }
+  ]
+}
